#70 feat : Nft 교환하기 상세페이지 API 연결

+++ app/src/main/java/com/aoztg/greengrim/data/remote/NftAPI.kt
@@ -62,7 +62,7 @@ interface NftAPI {
     @GET("/visitor/nfts/stock/refresh")
     suspend fun getNftForExchangeRefresh(
         @Query("grade") grade : String,
-        @Query("nftList") nftList : List<Int>
+        @Query("nftList") nftList : List<Long>
     ): Response<NftSimpleResponse>
 
 }
